AvdP - Improved stuff
Battery eating is better now, but even though the message is sent,
OpenCog still thinks batteries persist. This may be because we're not
removing the battery gameobject in Unity...so the next perception cycle
just reports to OpenCog that it still exists. If you keep recreating
batteries it works ok....